refactor: 중복 코드 삭제 및 폴더 이름 변경 #93
