+++ src/app/concert/form/[id]/_shared/components/input/form-input.module.scss
@@ -32,35 +32,10 @@
       display: flex;
       width: 100%;
       padding: 12px 0px;
-      align-items: flex-start;
-      border-bottom: 1px solid var(--lineColor-main);
-    }
-
-    .input_price,
-    .input_seat {
-      @include text-sm;
-      color: var(--textColor-less);
-
-      display: flex;
-      width: 100%;
-      padding: 12px 16px;
-      align-items: center;
-      border-radius: 12px;
-      background: var(--gray-1);
-      gap: 10px;
-      flex-shrink: 0;
-      border: none;
-    }
-
-    .select {
-      justify-content: space-between;
-    }
-
-    .textarea {
       height: 182px;
       align-items: flex-start;
       gap: 10px;
-
+      border-bottom: 1px solid var(--lineColor-main);
       outline: none;
 
       &:focus {
@@ -81,10 +56,7 @@
           @include text-sm;
           color: var(--textColor-less);
           text-decoration-line: underline;
-          // text-decoration-style: solid;
-          // text-decoration-skip-ink: auto;
-          // text-decoration-thickness: auto;
-          // text-underline-offset: auto;
+
           text-underline-position: from-font;
           cursor: pointer;
         }
@@ -119,14 +91,6 @@
         color: var(--textColor-main);
         text-align: center;
         min-width: 44px;
-      }
-
-      .input_price {
-        min-width: 109px;
-      }
-
-      .input_seat {
-        min-width: 137px;
       }
 
       & > :nth-child(1) {
